Add stop search feature
Stop search allows user to drag a map around and find nearby bus stops.
Clicking on a stop icon will show the user the upcoming bus arrivals for
that stop. The stop search allows users to set the maps position to
either their current position or the location of an intersection or
address.